refactor: answer util delete
