+++ src/component/question/basic-question.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import styled from 'styled-components';
 
 import GradientBox from '@/component/common/GradientBox';
