+++ src/component/result/IDCard/common/character-profile.tsx
@@ -7,7 +7,7 @@ import {
   ProfileTagContainerStyled,
   ProfileTagStyled,
 } from '@/component/result/IDCard/styles';
-import type { CharacterReturnType } from '@/utils/answer';
+import type { CharacterReturnType } from '@/types/character';
 
 export default function CharacterProfile({
   character,
