+++ src/types/character.ts
@@ -1,2 +1,7 @@
 export type CharacterType = 'lanny' | 'lumy' | 'cat';
 export type CharacterColorType = 'green' | 'yellow' | 'purple' | 'chee' | 'da';
+
+export type CharacterReturnType = {
+  name: string;
+  image: string;
+};
