+++ src/utils/character/image.ts
@@ -1,3 +1,5 @@
+import type { CharacterColorType, CharacterType } from '@/types/character';
+
 export const getCharacterImageSize = (
   character: string,
 ): {
@@ -26,4 +28,41 @@ export const getCharacterImageSize = (
         height: 191,
       };
   }
+};
+
+export const mappingColorValue = {
+  '1': 'blue',
+  '2': 'purple',
+  '3': 'yellow',
+  blue: '1',
+  purple: '2',
+  yellow: '3',
+};
+
+export const getColorLumiUrl = (value: string) => {
+  if (value === mappingColorValue['blue']) return '/characters/blue_lumy.png';
+  if (value === mappingColorValue['purple'])
+    return '/characters/purple_lumy.png';
+  if (value === mappingColorValue['yellow'])
+    return '/characters/yellow_lumy.png';
+  return '/characters/lumi.png';
+};
+
+function capitalize(str: string) {
+  if (!str) return str;
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
+
+export const getCharacterImageUrl = ({
+  color,
+  char,
+}: {
+  color: CharacterColorType;
+  char: CharacterType;
+}) => {
+  const data = {
+    name: `${capitalize(color)}_${capitalize(char)}`,
+    image: `/characters/${color}_${char}.png`,
+  };
+  return data;
 };
